feat: data table pulling score data from the api
